move model class to own package
